Supports multiple animations, Tanks now can vary stats

+++ TankGame/GameInfo/BoardState.cs
@@ -51,7 +51,7 @@ namespace TankGame.GameInfo
             //clone the players in the player list
             for (int i = 0; i < curBoardState.playerList.Count; i++)
             {
-                @newBoardState.playerList.Add(new Player(curBoardState.playerList[i].AP, curBoardState.playerList[i].inventory.sweeps));
+                @newBoardState.playerList.Add(new Player((int)curBoardState.playerList[i].AP, curBoardState.playerList[i].inventory.sweeps));
                 foreach (Tank tank in curBoardState.playerList[i].tanks)
                 {
                     @newBoardState.playerList[i].tanks.Add(Tank.Clone(tank));
